refs #55586 - refactored helper for deallocating cancelled processes

+++ src/Zmsdb/Helper/EventLogCleanUpByCron.php
@@ -11,9 +11,29 @@ use BO\Zmsdb\EventLog as EventLogRepository;
 
 class EventLogCleanUpByCron
 {
-    public static function startProcessing()
+    protected $verbose = false;
+    
+    public function __construct($verbose = false)
+    {
+        if ($verbose) {
+            $this->verbose = true;
+            $this->log("INFO: Delete old eventlog entries");
+        }
+    }
+
+    protected function log($message)
+    {
+        if ($this->verbose) {
+            error_log($message);
+        }
+    }
+
+    public static function startProcessing($commit = false)
     {
         $eventLogRepo  = new EventLogRepository();
-        $eventLogRepo->deleteOutdated();
+        if ($commit) {
+            $eventLogRepo->deleteOutdated();
+        }
+        
     }
 }
